refactor: extract shared AtomicWriteFile to pkg/fsutil
Deduplicate atomic write logic from cache.go and kubeconfig.go into
a shared fsutil.AtomicWriteFile utility to fix jscpd copy-paste
detection.
