+++ pkg/fsutil/atomic.go
@@ -1,0 +1,58 @@
+package fsutil
+
+import (
+	"fmt"
+	"os"
+	"path/filepath"
+	"runtime"
+)
+
+// AtomicWriteFile writes data to a temp file in the same directory and
+// renames it to the target path, ensuring an all-or-nothing write.
+// On Windows, if the rename fails because the destination exists, it
+// removes the target and retries once.
+func AtomicWriteFile(path string, data []byte, perm os.FileMode) error {
+	dir := filepath.Dir(path)
+
+	tmp, err := os.CreateTemp(dir, ".atomic-*.tmp")
+	if err != nil {
+		return fmt.Errorf("create temp file: %w", err)
+	}
+
+	tmpPath := tmp.Name()
+
+	defer func() {
+		_ = os.Remove(tmpPath)
+	}()
+
+	if chmodErr := os.Chmod(tmpPath, perm); chmodErr != nil {
+		_ = tmp.Close()
+
+		return fmt.Errorf("set permissions: %w", chmodErr)
+	}
+
+	if _, writeErr := tmp.Write(data); writeErr != nil {
+		_ = tmp.Close()
+
+		return fmt.Errorf("write data: %w", writeErr)
+	}
+
+	if closeErr := tmp.Close(); closeErr != nil {
+		return fmt.Errorf("close temp file: %w", closeErr)
+	}
+
+	renameErr := os.Rename(tmpPath, path)
+	if renameErr != nil && runtime.GOOS == "windows" {
+		if _, statErr := os.Stat(path); statErr == nil {
+			_ = os.Remove(path)
+
+			renameErr = os.Rename(tmpPath, path)
+		}
+	}
+
+	if renameErr != nil {
+		return fmt.Errorf("rename temp file: %w", renameErr)
+	}
+
+	return nil
+}
