+++ pkg/svc/oidc/cache.go
@@ -7,9 +7,10 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"runtime"
 	"sort"
 	"time"
+
+	"github.com/devantler-tech/ksail/v7/pkg/fsutil"
 )
 
 const (
@@ -102,57 +103,9 @@ func SaveCachedToken(cacheDir, key string, token *TokenResult) error {
 
 	cachePath := filepath.Join(cacheDir, key+".json")
 
-	writeErr := atomicWriteFile(cachePath, data, cacheFilePerm)
+	writeErr := fsutil.AtomicWriteFile(cachePath, data, cacheFilePerm)
 	if writeErr != nil {
 		return fmt.Errorf("failed to write cache file: %w", writeErr)
-	}
-
-	return nil
-}
-
-// atomicWriteFile writes data to a temp file in the same directory and
-// renames it to the target path, ensuring an all-or-nothing write.
-func atomicWriteFile(path string, data []byte, perm os.FileMode) error {
-	dir := filepath.Dir(path)
-
-	tmp, err := os.CreateTemp(dir, ".oidc-cache-*.tmp")
-	if err != nil {
-		return fmt.Errorf("create temp file: %w", err)
-	}
-
-	tmpPath := tmp.Name()
-
-	defer func() {
-		_ = os.Remove(tmpPath)
-	}()
-
-	if chmodErr := os.Chmod(tmpPath, perm); chmodErr != nil {
-		_ = tmp.Close()
-
-		return fmt.Errorf("set permissions: %w", chmodErr)
-	}
-
-	if _, writeErr := tmp.Write(data); writeErr != nil {
-		_ = tmp.Close()
-
-		return fmt.Errorf("write data: %w", writeErr)
-	}
-
-	if closeErr := tmp.Close(); closeErr != nil {
-		return fmt.Errorf("close temp file: %w", closeErr)
-	}
-
-	renameErr := os.Rename(tmpPath, path)
-	if renameErr != nil && runtime.GOOS == "windows" {
-		if _, statErr := os.Stat(path); statErr == nil {
-			_ = os.Remove(path)
-
-			renameErr = os.Rename(tmpPath, path)
-		}
-	}
-
-	if renameErr != nil {
-		return fmt.Errorf("rename temp file: %w", renameErr)
 	}
 
 	return nil
